Funcionamiento de condicionales agregada: No se pueden hacer residuos o divisiones entre cero. Solo queda por hacer el HILO y tentativamente la acotación de getlines

+++ proceso.h
@@ -2,6 +2,8 @@
 #define PROCESO_H
 
 #include <string>
+#include <sstream>
+#include <iomanip>
 
 class Proceso {
     public:
@@ -9,22 +11,22 @@ class Proceso {
         virtual ~Proceso();
         void setId(const int & id);
         void setProgramador(const std::string&programador);
-        void setN1(const int & n1);
-        void setN2(const int & n2);
+        void setN1(const float & n1);
+        void setN2(const float & n2);
         void setOperador(const std::string&operador);
         void setTiempoEstimado(const int & tiempoEstimado);
         int getId();
         std::string getProgramador();
-        int getN1();
-        int getN2();
+        float getN1();
+        float getN2();
         std::string getOperador();
         int getTiempoEstimado();
         std::string getOperacion();
     private:
         int id;
         std::string programador;
-        int n1;
-        int n2;
+        float n1;
+        float n2;
         std::string operador;
         int tiempoEstimado;
 };
